requirements.txt
pandas
numpy
matplotlib
alpaca-py
python-dotenv
scipy
pytz
nest-asyncio